requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.13.0
google-cloud-pubsub==2.18.4
pandas==2.1.3
